feat: deposit account transaction resource, service + tests

+++ backend/src/main/java/de/stecknitz/backend/core/repository/DepositAccountRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DepositAccountRepository extends JpaRepository<DepositAccount, Integer> {
+public interface DepositAccountRepository extends JpaRepository<DepositAccount, Long> {
 }
